requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
sqlmodel==0.0.21
httpx==0.27.0
python-dotenv==1.0.1
APScheduler==3.10.4
Jinja2==3.1.4
python-multipart==0.0.9